app/requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
defusedxml==0.7.1
diff-match-patch==20200713
Django>=3.2.12,<3.3
psycopg2>=2.8.6,<2.9
django-cors-headers==3.13.0
django-import-export==2.8.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
google-api-core==2.8.2
google-api-python-client==2.55.0
# google-auth==2.9.1
# google-auth-httplib2==0.1.0
# google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.3
import-export==0.2.67.dev6
MarkupPy==1.14
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.10
protobuf==4.21.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.1.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
timeloop==1.0.2
tqdm==4.64.0
uritemplate==4.1.1
urllib3==1.26.11
xlrd==2.0.1
xlwt==1.3.0
